Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -15,7 +15,7 @@ int main(){
     }
     else if(operator=='/'){
         if(b!=0){
-            printf("%d",a,b,a/b);
+            printf("%d",a,b,(float)a/b);
         else{
             printf("Error");
         }
